refactor and init Board modules
