+++ nim.cabal
@@ -38,7 +38,7 @@ executable nim
     main-is:            Main.hs
 
     -- Modules included in this executable, other than Main.
-    other-modules:      Nim 
+    other-modules:      Nim, Board 
 
     hs-source-dirs:     app, src
